+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking changes
 
+* Package Kibana compatiblity version is changed to `"kibana": { "max": "1.2.3"}` [#](https://github.com/elastic/integrations-registry/pull/)
+
+
 ### Bugfixes
 
 * Change empty /search API output from `null` to `[]`. [#111](https://github.com/elastic/integrations-registry/pull/111)